Update FileUploadValidator.java
In this updated code, the filename validation is performed using a regular expression pattern. If the filename is invalid, the file is rejected and moved to the quarantine folder. The filepath is not disclosed in any of the output messages. Exceptions are handled securely by printing a generic error message and stack trace is not printed.